+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+### Added
+
+- Added comment wrapping document and preamble comments.
+
 ## 0.4.0 - 08-12-2025
 
 ### Added